Optimize the code structure
