+++ vispy_isosurface/__init__.py
@@ -1,6 +1,6 @@
 __author__ = 'penny'
 
 def setup():
-    from .iso_glueViewer import GlueIsoVispyViewer
+    from .iso_glue_viewer import GlueIsoVispyViewer
     from glue.config import qt_client
     qt_client.add(GlueIsoVispyViewer)
