--- vispy_isosurface/iso_gluePlugin.py
@@ -1,6 +1,0 @@
-__author__ = 'penny'
-
-def setup():
-    from .iso_glueViewer import GlueIsoVispyViewer
-    from glue.config import qt_client
-    qt_client.add(GlueIsoVispyViewer)
