+++ vispy_isosurface/tests/test_iso_glueViewer.py
@@ -1,5 +1,5 @@
 from glue.qt import get_qapp
-from ..iso_glueViewer import GlueIsoVispyViewer
+from ..iso_glue_viewer import GlueIsoVispyViewer
 from glue.config import qt_client
 
 def test_viewer():
