+++ vispy_volume/__init__.py
@@ -2,6 +2,6 @@ __author__ = 'penny'
 
 
 def setup():
-    from .vol_glueViewer import GlueVispyViewer
+    from .vol_glue_viewer import GlueVispyViewer
     from glue.config import qt_client
     qt_client.add(GlueVispyViewer)
